refactoring orderproduct

+++ src/main/java/com/orxan/sweetstorerest/service/serviceimple/ProductServiceImpl.java
@@ -1,39 +1,50 @@
 package com.orxan.sweetstorerest.service.serviceimple;
 
 import com.orxan.sweetstorerest.exceptions.InvalidProductException;
+import com.orxan.sweetstorerest.exceptions.ResourceNotFoundException;
 import com.orxan.sweetstorerest.model.Product;
 import com.orxan.sweetstorerest.repository.daoimpl.ProductDaoImpl;
 import com.orxan.sweetstorerest.service.ProductService;
-import com.orxan.sweetstorerest.util.LoadPropertyUtil;
 import com.orxan.sweetstorerest.util.NumberUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.Properties;
 
 @Service
+@PropertySource("classpath:properties/errors.properties")
 public class ProductServiceImpl implements ProductService {
 
     @Autowired
     private ProductDaoImpl productDao;
-    private Properties errorProperties;
-    private String ERROR_PROPERTIES_URL="properties/errors.properties";
-    private static Map<String, Map<Boolean, List<String>>> validation;
-
-    public ProductServiceImpl() {
-        this.errorProperties = LoadPropertyUtil.loadPropertiesFile(ERROR_PROPERTIES_URL);
-    }
+    @Value("${error.product.nameSize}")
+    private String nameSize;
+    @Value("error.product.nullName")
+    private String nullName;
+    @Value("error.product.negativeQuantity")
+    private String negativeQuantity;
+    @Value("error.product.maxQuantity")
+    private String maxQuantity;
+    @Value("error.product.invalidNumber")
+    private String invalidNumber;
+    @Value("error.product.negativePrice")
+    private String negativePrice;
+    @Value("error.product.maxPrice")
+    private String maxPrice;
 
     @Override
     public List<Product> getProductList(int pageIndex, int rowsPerPage) {
         int totalCount= getTotalCountOfProduct();
         int fromIndex=pageIndex*rowsPerPage;
         int toIndex=Math.min(fromIndex+rowsPerPage,totalCount);
-        return productDao.getProductList(fromIndex,toIndex);
+        List<Product> productList=productDao.getProductList(fromIndex,toIndex);
+        if (!productList.isEmpty()) {
+            return productList;
+        } else throw new ResourceNotFoundException("No products found");
     }
 
     @Override
@@ -50,11 +61,15 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public Product updateProduct(Product product, int oldProductId) {
         product.setUpdateDate(LocalDateTime.now());
-        List<String> errorList= isProductValid(product);
-        if (errorList.isEmpty()) {
-            return productDao.updateProduct(product, oldProductId);
-        }else {
-            throw new InvalidProductException(errorList);
+        if (productDao.isProductExist(oldProductId)) {
+            List<String> errorList = isProductValid(product);
+            if (errorList.isEmpty()) {
+                return productDao.updateProduct(product, oldProductId);
+            } else {
+                throw new InvalidProductException(errorList);
+            }
+        } else {
+            throw new ResourceNotFoundException("Product not found. Id="+oldProductId);
         }
     }
 
@@ -65,31 +80,31 @@ public class ProductServiceImpl implements ProductService {
             if (product.getName() != null) {
                 String name = product.getName();
                 if (name.trim().length() < 2) {
-                    errorList.add(errorProperties.getProperty("nameSize"));
+                    errorList.add(nameSize);
                 } else {
                     product.setName(renameProduct(product.getName()));
                 }
             } else {
-                errorList.add(errorProperties.getProperty("nullName"));
+                errorList.add(nullName);
             }
 
             if (product.getQuantity() < 0) {
-                errorList.add(errorProperties.getProperty("negativeQuantity"));
+                errorList.add(negativeQuantity);
             }
             if (product.getQuantity() > 1000) {
-                errorList.add(errorProperties.getProperty("maxQuantity"));
+                errorList.add(maxQuantity);
             }
             if (!NumberUtils.isNumberInteger(String.valueOf(product.getQuantity()))) {
-                errorList.add(errorProperties.getProperty("invalidNumber"));
+                errorList.add(invalidNumber);
             }
             if (product.getPrice() < 0) {
-                errorList.add(errorProperties.getProperty("negativePrice"));
+                errorList.add(negativePrice);
             }
             if (product.getPrice() > 1000) {
-                errorList.add(errorProperties.getProperty("maxPrice"));
+                errorList.add(maxPrice);
             }
             if (!NumberUtils.isNumberFloat(String.valueOf(product.getPrice()))) {
-                errorList.add(errorProperties.getProperty("invalidNumber"));
+                errorList.add(invalidNumber);
             }
         }
         return errorList;
@@ -97,13 +112,19 @@ public class ProductServiceImpl implements ProductService {
 
     @Override
     public boolean deleteProductByID(int id) {
-        return productDao.deleteProductById(id);
+       boolean exist =productDao.isProductExist(id);
+       if (exist) {
+           productDao.deleteProductById(id);
+           return true;
+       } else throw new ResourceNotFoundException("Product not found. Id="+id);
     }
 
     @Override
     public Product getProductById(int id) {
         Product product=productDao.getProductById(id);
-       return product;
+        if (product!=null) {
+            return product;
+        } else throw new ResourceNotFoundException("Product not found. id="+id);
     }
 
     @Override
